improved on init scripts
git-svn-id: https://bigstreams.googlecode.com/svn/trunk@75 e9027e87-718f-2365-ccfc-279084edcb2a

+++ collector/src/main/resources/etc/init.d/streams-collector.sh
@@ -3,8 +3,9 @@ export PATH
 
 [ -f /etc/sysconfig/streams-collector ] && . /etc/sysconfig/streams-collector
 lockfile=${LOCKFILE-/var/lock/subsys/streams-collector}
-pidfile=${PIDFILE-/var/run/streams-collector.pid}
 streamsd=${STREAMSD-/opt/streams-collector/bin/streams.sh}
+REGEX="org.streams.collector.main.Main -start collector"
+
 RETVAL=0
 
 # Source function library.
@@ -12,20 +13,42 @@ RETVAL=0
 
 
 start() {
+
+   pid=`pgrep -f "$REGEX"`
+
+   if [ -n "$pid" ]; then
+    echo "The Streams collector is already running"
+    RETVAL=2
+   else
+
     echo -n $"Starting streams collector: "
         daemon $streamsd -start collector &> /dev/null &
     RETVAL=$?
     echo
         [ $RETVAL = 0 ] && touch ${lockfile}
-        return $RETVAL
+
+   fi
+
+   return $RETVAL
+
 }
 
 stop() {
-    echo -n $"Stopping streams: "
+
+    pid=`pgrep -f "$REGEX"`
+    if [ -n "$pid" ]; then
+      echo -n $"Stopping streams: "
         $streamsd -stop collector
-    RETVAL=$?
-    echo
-    [ $RETVAL = 0 ] && rm -f ${lockfile} ${pidfile}
+      RETVAL=$?
+      echo
+       [ $RETVAL = 0 ] && rm -f ${lockfile} ${pidfile}
+    else
+       echo "No Streams collector instance is running" 
+       RETVAL=2
+    fi
+
+    return $RETVAL
+
 }
 
 restart() {
@@ -33,9 +56,19 @@ restart() {
     start
 }
 
-rh_status() {
-    status | grep -q -- '-p' 2>/dev/null && statusopts="-p $pidfile"
-    status $statusopts $streamsd
+status() {
+
+   pid=`pgrep -f "$REGEX"`
+   if [ -n "$pid" ]; then
+      echo "Streams collector is running"
+      echo "$pid"
+      RETVAL=1
+   else
+       echo "Streams collector instance is stopped" 
+       RETVAL=2
+   fi
+
+   return $RETVAL
 }
 
 case "$1" in
@@ -46,17 +79,14 @@ case "$1" in
     stop
     ;;
   restart)
-    restart
-    ;;
-  condrestart|try-restart)
-    rh_status_q || exit 0
-    restart
+    stop
+    start
     ;;
   status)
-        rh_status
+        status
     ;;
   *)
-    echo $"Usage: $0 {start|stop|status|restart|condrestart}"
+    echo $"Usage: $0 {start|stop|status|restart}"
     exit 1
 esac
